+++ pkg/http_server/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.9.5
+
+* Removed the decoding of HTML entity values (in the form &#xxxxx;) for
+  values when parsing multipart/form-post requests.
+
 # 0.9.4
 
 * Fixed bugs in the handling of the Range header